// src/app/data.interface.ts
export interface Group {
  value: string;
  viewValue: string;
}

export interface ParkingCategory {
  value: string;
  viewValue: string;
}

export interface ParkingType {
  value: string;
  viewValue: string;
}

export interface PricingType {
  value: string;
  viewValue: string;
}

export interface Airport {
  value: string;
  viewValue: string;
}

export interface Dates {
  value: string;
  viewValue: string;
}

export interface Hour {
  value: string;
  viewValue: string;
}

export interface Time {
  value: string;
  viewValue: string;
}

export interface Minute {
  value: string;
  viewValue: string;
}

export interface OtherPricingType {
  value: string;
  viewValue: string;
}

export interface CheckoutType {
  value: string;
  viewValue: string;
}